fix(test): run Contacts .limited coverage in the live permissions suite
runPermissionManagerTests() is skipped (checkAll hangs in CLI). CI measured
3926 passed / 0 failed against FLOOR 3927 because the .limited test never
ran. Move it to PermissionsModuleTests so the seventh new test is counted.
