+++ TheBridgeTests/PermissionManagerTests.swift
@@ -9,7 +9,6 @@
 // that detection methods don't crash or block the main thread.
 
 import Foundation
-import Contacts
 import TheBridgeLib
 
 // MARK: - PermissionManager Tests
@@ -66,20 +65,6 @@ func runPermissionManagerTests() async {
         try expect(PermissionManager.Grant.fullDiskAccess.isAutoGrantable == false)
         try expect(PermissionManager.Grant.reminders.isAutoGrantable == false)
         try expect(PermissionManager.Grant.calendar.isAutoGrantable == false)
-    }
-
-    await test("Contacts .limited authorization is sufficient") {
-        // The Swift overlay marks CNAuthorizationStatus.limited unavailable
-        // on macOS (API_UNAVAILABLE(macos)), even though Sequoia Selected
-        // Contacts reports rawValue 4 at runtime. Construct via rawValue so
-        // the test compiles and still covers PermissionManager's .limited arm.
-        let limited = CNAuthorizationStatus(rawValue: 4)!
-        try expect(PermissionManager.isContactsAuthorizationSufficient(limited),
-                   "Selected Contacts (.limited) must be enough for contacts_* tools")
-        try expect(PermissionManager.isContactsAuthorizationSufficient(.authorized))
-        try expect(!PermissionManager.isContactsAuthorizationSufficient(.denied))
-        try expect(!PermissionManager.isContactsAuthorizationSufficient(.restricted))
-        try expect(!PermissionManager.isContactsAuthorizationSufficient(.notDetermined))
     }
 
     await test("GrantStatus equality works correctly") {
