+++ scripts/test-floor-gate-history.md
@@ -2460,4 +2460,4 @@ set -euo pipefail
 - 2026-09-02 · #189 inbound /health counters · `3852 → 3854` (+2): expose `/mcp` inbound count on `/health` for Notion connect flake diagnosis. Measured 3854 passed, 0 failed on feat/issue-189-inbound-health after rebase onto main (post-#201).
 - 2026-09-02 · #232–#235 registry Wave 1 · `3854 → 3859` (+5): binder follows property id across rename; writes keyed by decoded id; cache revalidates `lastEditedTime`; Notion URL-or-UUID including `app.notion.com`. Measured 3859 passed, 0 failed on wave1-registry-234.
 - 2026-09-02 · Notion REST Wave 2 · `3859 → 3890` (+31): view query/delete, page trash/restore, search `in_trash`, comment PATCH/DELETE, async tasks, templates XOR children, meeting-notes query, opt-in multi_part upload (#225–#231, #236–#237). Measured 3890 passed, 0 failed on wave2-notion-rest.
-- 2026-09-03 · hotfix #238 Safari Cmd+V · `3920 → 3927` (+7): Safari/WebKit prefers paste; leftover ⌃⌥⇧ HID key-up before Cmd+V (Command held); Command-held UCKeyTranslate for V; pasteboard restore-on-throw; gh_pr_review drops `comment`; keyboard_type copy drops Input Monitoring claim; Contacts `.limited` sufficient. Expected +7 green vs 4.0.7 floor 3920; confirm on macOS CI. FLOOR 3920 -> 3927.
+- 2026-09-03 · hotfix #238 Safari Cmd+V · `3920 → 3927` (+7): Safari/WebKit prefers paste; leftover ⌃⌥⇧ HID key-up before Cmd+V (Command held); Command-held UCKeyTranslate for V; pasteboard restore-on-throw; gh_pr_review drops `comment`; keyboard_type copy drops Input Monitoring claim; Contacts `.limited` sufficient (live runner is PermissionsModuleTests — `runPermissionManagerTests` is skipped because `checkAll()` hangs). CI measured 3926 passed / 0 failed on `1f18ed4` before moving the Contacts test; floor 3927 after that move.
